feature/artistManagerScheduling

+++ database/migrations/2018_11_18_150446_create_applications_table.php
@@ -13,7 +13,7 @@ class CreateApplicationsTable extends Migration {
 	public function up() {
 		Schema::create('applications', function (Blueprint $table) {
 			$table->bigIncrements('id');
-			$table->integer('kitchen_id')->unsigned();
+			$table->bigInteger('kitchen_id')->unsigned();
 			$table->integer('number')->nullable();
 			$table->string('status');
 			$table->year('year');
