+++ database/migrations/2019_01_22_223903_create_worker_photos_table.php
@@ -13,7 +13,7 @@ class CreateWorkerPhotosTable extends Migration {
 	public function up() {
 		Schema::create('worker_photos', function (Blueprint $table) {
 			$table->bigIncrements('id');
-			$table->integer('worker_id')->unsigned();
+			$table->bigInteger('worker_id')->unsigned();
 			$table->string('file');
 			$table->timestamps();
