+++ database/migrations/2019_01_25_133738_create_shifts_table.php
@@ -13,7 +13,7 @@ class CreateShiftsTable extends Migration {
 	public function up() {
 		Schema::create('shifts', function (Blueprint $table) {
 			$table->bigIncrements('id');
-			$table->integer('workplace_id')->unsigned();
+			$table->bigInteger('workplace_id')->unsigned();
 			$table->date('date');
 			$table->integer('hours');
 			$table->boolean('closed')->default(false);
